requirements.txt
langchain
openai
sentence-transformers
streamlit
streamlit-chat
langgraph
psycopg2
chromadb

langchain_google_community
langchain-google-community[drive]
langchain-chroma
unstructured[pdf]
pytesseract
pillow
easyocr
transformers
torch